Enable truffle to use mixed SVE and neon
The match component is faster with SVE but the mask one is faster
with neon. This patch enable the use of the neon sve bridge that
allow converting vector from one to the other, thus using the best
implementation for each truffle component.
This hybrid version speeds truffle by about 10% on the microbenchmark.
This is available with Clang 14/gcc 14. Otherwise it default to all
neon/SVE

Signed-off-by: Yoan Picchi <yoan.picchi@arm.com>